fix(backup): unpredictable uploads
Because tar does not append file into a tar.gz file
it was returning to an error. Since it is not explicitly
handled, the rest of the backup candidates were dropped
by the function since it returns with an error.

Now it just prints the reason and continue with the next element
from the Vec.

Signed-off-by: Taylan Dogan <git@taylandogan.info>